A fix for the top rated link not working. Had to repeat code.
